new feature to allow settings of default preferences on install based on property file (needed for CSP) added new preference to turn off automatic survey checks (to conserve data) new preference to control when remote exceptions are uploaded

+++ WFPMapping/device/survey/src/com/gallatinsystems/survey/device/service/SurveyDownloadService.java
@@ -122,50 +122,56 @@ public class SurveyDownloadService extends Service {
 				} else {
 					serverBase = props.getProperty(ConstantUtil.SERVER_BASE);
 				}
-				ArrayList<Survey> surveys = null;
 
-				if (surveyId != null && surveyId.trim().length() > 0) {
-					surveys = getSurveyHeader(serverBase, surveyId);
-					if (surveys != null && surveys.size() > 0) {
-						// if we already have the survey, delete it first
-						databaseAdaptor.deleteSurvey(surveyId.trim(), true);
+				int surveyCheckOption = Integer.parseInt(databaseAdaptor
+						.findPreference(ConstantUtil.CHECK_FOR_SURVEYS));
+				if (canDownload(surveyCheckOption)) {
+					ArrayList<Survey> surveys = null;
+
+					if (surveyId != null && surveyId.trim().length() > 0) {
+						surveys = getSurveyHeader(serverBase, surveyId);
+						if (surveys != null && surveys.size() > 0) {
+							// if we already have the survey, delete it first
+							databaseAdaptor.deleteSurvey(surveyId.trim(), true);
+						}
+					} else {
+						surveys = checkForSurveys(serverBase);
 					}
-				} else {
-					surveys = checkForSurveys(serverBase);
-				}
-				if (surveys != null && surveys.size() > 0) {
-					// if there are surveys for this device, see if we need them
-					surveys = databaseAdaptor.checkSurveyVersions(surveys);
-					int updateCount = 0;
 					if (surveys != null && surveys.size() > 0) {
-						for (int i = 0; i < surveys.size(); i++) {
-							Survey survey = surveys.get(i);
-							try {
-								if (downloadSurvey(serverBase, survey)) {
-									databaseAdaptor.saveSurvey(survey);
-									downloadHelp(survey, precacheOption);
-									updateCount++;
+						// if there are surveys for this device, see if we need
+						// them
+						surveys = databaseAdaptor.checkSurveyVersions(surveys);
+						int updateCount = 0;
+						if (surveys != null && surveys.size() > 0) {
+							for (int i = 0; i < surveys.size(); i++) {
+								Survey survey = surveys.get(i);
+								try {
+									if (downloadSurvey(serverBase, survey)) {
+										databaseAdaptor.saveSurvey(survey);
+										downloadHelp(survey, precacheOption);
+										updateCount++;
+									}
+								} catch (Exception e) {
+									Log.e(TAG, "Could not download survey", e);
+									PersistentUncaughtExceptionHandler
+											.recordException(e);
 								}
-							} catch (Exception e) {
-								Log.e(TAG, "Could not download survey", e);
-								PersistentUncaughtExceptionHandler
-										.recordException(e);
+							}
+							if (updateCount > 0) {
+								fireNotification(updateCount);
 							}
 						}
-						if (updateCount > 0) {
-							fireNotification(updateCount);
-						}
 					}
-				}
 
-				// now check if any previously downloaded surveys still need
-				// don't have their help media pre-cached
-				if (canDownload(precacheOption)) {
-					surveys = databaseAdaptor.listSurveys(null);
-					if (surveys != null) {
-						for (int i = 0; i < surveys.size(); i++) {
-							if (!surveys.get(i).isHelpDownloaded()) {
-								downloadHelp(surveys.get(i), precacheOption);
+					// now check if any previously downloaded surveys still need
+					// don't have their help media pre-cached
+					if (canDownload(precacheOption)) {
+						surveys = databaseAdaptor.listSurveys(null);
+						if (surveys != null) {
+							for (int i = 0; i < surveys.size(); i++) {
+								if (!surveys.get(i).isHelpDownloaded()) {
+									downloadHelp(surveys.get(i), precacheOption);
+								}
 							}
 						}
 					}
@@ -491,11 +497,12 @@ public class SurveyDownloadService extends Service {
 		boolean ok = false;
 		if (precacheOptionIndex > -1
 				&& ConstantUtil.PRECACHE_WIFI_ONLY_IDX == precacheOptionIndex) {
-			
+
 			ok = StatusUtil.hasDataConnection(this, true);
 		} else {
 			ok = StatusUtil.hasDataConnection(this, false);
 		}
 		return ok;
 	}
+
 }
